[DESKTOP/API] Rename "register" into "sign up"
This commit renames the "register" route and files into "sign up".

closes #74
