+++ projects/desktop/src/components/atoms/common/background/Background.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import background from '../../../../assets/img/background.png'
 
 /**
- * Main background (clouds) on the login & register page
+ * Main background (clouds) on the login & signup page
  */
 export const Background = styled.div`
     background-image: url(${background});
