+++ projects/desktop/src/components/atoms/signup/SignUpBox.tsx
@@ -4,9 +4,9 @@ import { Box as MaterialBox } from '@material-ui/core'
 import { colors } from '../../../constants'
 
 /**
- * Main box on the register page
+ * Main box on the signup page
  */
-export const RegisterBox = styled(MaterialBox)`
+export const SignUpBox = styled(MaterialBox)`
     background-color: ${colors.white};
     border-radius: 8px;
     box-shadow: 4px 8px 32px 0 rgba(0,0,0,0.10);
