+++ projects/desktop/src/components/atoms/signup/SubmitButton.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { GradientButton } from '../singleForm/GradientButton'
 
 /**
- * Big register button
+ * Big signup button
  */
 export const SubmitButton = styled(GradientButton)`
   margin-top: 30px !important;
